tour 작성 권한 (로그인) 삭제/수정 권한(에러)

+++ be/src/main/java/com/example/be/controller/tour/TourController.java
@@ -4,6 +4,7 @@ import com.example.be.dto.tour.Tour;
 import com.example.be.service.tour.TourService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,27 +24,42 @@ public class TourController {
   }
 
   @PutMapping("update")
+  @PreAuthorize("isAuthenticated()")
   public ResponseEntity<Map<String, Object>> update(
-          Tour tour,
+          Tour tour, Authentication authentication,
           @RequestParam(value = "removeFiles[]", required = false) List<String> removeFiles,
           @RequestParam(value = "uploadFiles[]", required = false) MultipartFile[] uploadFiles) {
-    if (service.update(tour, removeFiles, uploadFiles)) {
-      return ResponseEntity.ok(Map.of("message",
-              Map.of("type", "success", "text", "상품 수정 완료")));
+    if (service.hasAccess(tour.getId(), authentication)) {
+      if (service.update(tour, removeFiles, uploadFiles)) {
+        return ResponseEntity.ok(Map.of("message",
+                Map.of("type", "success", "text", "상품 수정 완료")));
+      } else {
+        return ResponseEntity.ok(Map.of("message",
+                Map.of("type", "warning", "text", "상품 수정 실패")));
+      }
     } else {
-      return ResponseEntity.ok(Map.of("message",
-              Map.of("type", "warning", "text", "상품 수정 실패")));
+      return ResponseEntity.status(403).body(
+              Map.of("message", Map.of("type", "error", "text", "수정 권한이 없습니다.")));
     }
   }
 
+
   @DeleteMapping("delete/{id}")
-  public ResponseEntity<Map<String, Object>> delete(@PathVariable int id) {
-    if (service.delete(id)) {
-      return ResponseEntity.ok().body(Map.of("msg",
-              Map.of("type", "success", "text", "상품 삭제")));
+  @PreAuthorize("isAuthenticated()")
+  public ResponseEntity<Map<String, Object>> delete(
+          @PathVariable int id, Authentication authentication) {
+
+    if (service.hasAccess(id, authentication)) {
+      if (service.delete(id)) {
+        return ResponseEntity.ok().body(Map.of("message",
+                Map.of("type", "success", "text", "상품 삭제")));
+      } else {
+        return ResponseEntity.status(400).body(Map.of("message",
+                Map.of("type", "warning", "text", "상품 수정 실패")));
+      }
     } else {
-      return ResponseEntity.status(400).body(Map.of("message",
-              Map.of("type", "warning", "text", "상품 수정 실패")));
+      return ResponseEntity.status(403).body(
+              Map.of("message", Map.of("type", "error", "text", "삭제 권한이 없습니다.")));
     }
   }
 
@@ -61,6 +77,7 @@ public class TourController {
   }
 
   @PostMapping("add")
+  @PreAuthorize("isAuthenticated()")
   public ResponseEntity<Map<String, Object>> add(
           Tour tour,
           @RequestParam(value = "files[]", required = false) MultipartFile[] files,
